Move parsing methods to trait Results in cleaner, more maintainable code
